# Faster `changebasis`