**Before this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> P1 = BSplineSpace{1}(KnotVector([1, 3, 5, 8]))
BSplineSpace{1, Int64, KnotVector{Int64}}(KnotVector([1, 3, 5, 8]))

julia> P2 = BSplineSpace{1}(KnotVector([1, 3, 5, 6, 8, 9]))
BSplineSpace{1, Int64, KnotVector{Int64}}(KnotVector([1, 3, 5, 6, 8, 9]))

julia> @benchmark changebasis(P1, P2)
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.898 μs … 350.734 μs  ┊ GC (min … max):  0.00% … 99.22%
 Time  (median):     2.059 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   2.501 μs ±  10.480 μs  ┊ GC (mean ± σ):  13.18% ±  3.12%

   ▄▆▇███▇▆▅▃▃▁▁▂▂▃▂▃▂▂▂▁▁▁▁                                  ▂
  ███████████████████████████████▇███▇███▇████████▇▆▇▆▇▇▆▆▆▆▄ █
  1.9 μs       Histogram: log(frequency) by time      3.47 μs <

 Memory estimate: 4.59 KiB, allocs estimate: 51.
```

**After this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> P1 = BSplineSpace{1}(KnotVector([1, 3, 5, 8]))
BSplineSpace{1, Int64, KnotVector{Int64}}(KnotVector([1, 3, 5, 8]))

julia> P2 = BSplineSpace{1}(KnotVector([1, 3, 5, 6, 8, 9]))
BSplineSpace{1, Int64, KnotVector{Int64}}(KnotVector([1, 3, 5, 6, 8, 9]))

julia> @benchmark changebasis(P1, P2)
BenchmarkTools.Trial: 10000 samples with 78 evaluations.
 Range (min … max):  829.256 ns … 39.370 μs  ┊ GC (min … max):  0.00% … 96.59%
 Time  (median):     865.090 ns              ┊ GC (median):     0.00%
 Time  (mean ± σ):   993.385 ns ±  1.955 μs  ┊ GC (mean ± σ):  10.73% ±  5.31%

   ▄▇██▇▆▄▃▂▁▁                                                 ▂
  ███████████████▇▇▇▇▇▇▇███▇█▇▆▆▆▆▄▄▄▃▅▅▅▆▇▆▅▆▇▇▇▆▇▇▇▆▆▅▄▆▅▆▆▆ █
  829 ns        Histogram: log(frequency) by time      1.29 μs <

 Memory estimate: 1.81 KiB, allocs estimate: 18.
```