@hyrodium 2023-03-08
Now the performance is much better:
```julia
julia> using BasicBSpline, BenchmarkTools

julia> P1 = BSplineSpace{1}(KnotVector([1, 3, 5, 8]))
BSplineSpace{1, Int64, KnotVector{Int64}}(KnotVector([1, 3, 5, 8]))

julia> P2 = BSplineSpace{1}(KnotVector([1, 3, 5, 6, 8, 9]))
BSplineSpace{1, Int64, KnotVector{Int64}}(KnotVector([1, 3, 5, 6, 8, 9]))

julia> @benchmark changebasis(P1, P2)
BenchmarkTools.Trial: 10000 samples with 173 evaluations.
 Range (min … max):  616.832 ns … 23.942 μs  ┊ GC (min … max):  0.00% … 96.58%
 Time  (median):     651.231 ns              ┊ GC (median):     0.00%
 Time  (mean ± σ):   778.790 ns ±  1.384 μs  ┊ GC (mean ± σ):  12.35% ±  6.72%

   ▅██▇▅▄▃▃▂▁▁▂▃▃▂▁                                            ▂
  █████████████████████████▇██▇▇▆▇▆▆▆▆▆▆▅▆▆▄▇▆▇▆▆▅▄▆▄▅▅▅▄▃▃▄▄▃ █
  617 ns        Histogram: log(frequency) by time      1.18 μs <

 Memory estimate: 1.28 KiB, allocs estimate: 14.
```